@@ -1,23 +1,23 @@
 using System.Threading.Tasks;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace Rabobank.TechnicalTest.GCOB.Tests.Controllers
+namespace Rabobank.TechnicalTest.GCOB.Tests.Controllers;
+
+[TestClass]
+public class CustomerControllerTest
 {
-    [TestClass]
-    public class CustomerControllerTest
-    {
-
-        [TestInitialize]
-        public void Initialize()
-        {
-
-        }
 
 
-        [TestMethod]
-        public async Task GivenHaveACustomer_AndICallAServiceToGetTheCustomer_ThenTheCustomerIsReturned()
-        {
 
-        }
-    }
+  [TestInitialize]
+  public void Initialize()
+  {
+
+  }
+
+  [TestMethod]
+  public async Task CustomerRepository_add_a_single_customer_then_retrieve_customer()
+  {
+
+  }
 }
