@@ -5,5 +5,5 @@ public class CustomerDto
   public int Id { get; set; }
   public string FirstName { get; set; }
   public string LastName { get; set; }
-  public int AddressId { get; set; }
+  public AddressDto Address { get; set; }
 }
