@@ -2,6 +2,6 @@
 {
   public abstract class Entity
   {
-    public int Id { get; set; }
+    public int Id { get; init; }
   }
 }
