@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using Rabobank.TechnicalTest.GCOB.Models.Entities.Abstract;
+using Rabobank.TechnicalTest.GCOB.Models.Exceptions;
 
 namespace Rabobank.TechnicalTest.GCOB.Models.Repositories.Abstract;
 
@@ -13,6 +14,8 @@ public abstract class InMemoryRepository<T> where T : Entity
 
   protected ConcurrentDictionary<int, T> Data { get; init; }
 
+  private readonly object _identityLock = new();
+
   protected InMemoryRepository(ILogger<InMemoryRepository<T>> logger)
   {
     Logger = logger;
@@ -20,34 +23,42 @@ public abstract class InMemoryRepository<T> where T : Entity
 
   protected abstract string EntityName { get; }
 
-  protected Task<int> GenerateIdentityAsync()
+  public Task<int> GenerateIdentityAsync()
   {
-    Logger.LogDebug("Generating Customer identity");
+    Logger.LogDebug($"Generating {EntityName} identity");
 
     return Task.Run(() =>
     {
-      if (Data.Count == 0) return 1;
+      lock (_identityLock)
+      {
+        CheckDataInstantiated();
 
-      var x = Data.Values.Max(c => c.Id);
-      return ++x;
+        if (Data.Count == 0) return 1;
+
+        var x = Data.Values.Max(c => c.Id);
+        return ++x;
+      }
     });
   }
   
   public Task<T> GetAsync(int identity)
   {
-    Logger.LogDebug($"FindMany Customers with identity {identity}");
+    Logger.LogDebug($"Find {EntityName} with identity {identity}");
 
-    if (!Data.ContainsKey(identity)) throw new Exception(identity.ToString());
-    Logger.LogDebug($"Found Customer with identity {identity}");
+    CheckDataInstantiated();
+
+    if (!Data.ContainsKey(identity)) throw new NoEntityFoundException($"No {EntityName} exists with identity {identity.ToString()}");
+    Logger.LogDebug($"Found {EntityName} with identity {identity}");
     return Task.FromResult(Data[identity]);
-
   }
 
   public Task InsertAsync(T entity)
   {
+    CheckDataInstantiated();
+
     if (Data.ContainsKey(entity.Id))
     {
-      throw new Exception(
+      throw new DuplicateEntityFoundException(
         $"Cannot insert {EntityName} with identity '{entity.Id}' as it already exists in the collection");
     }
 
@@ -58,9 +69,11 @@ public abstract class InMemoryRepository<T> where T : Entity
 
   public Task UpdateAsync(T entity)
   {
+    CheckDataInstantiated();
+
     if (!Data.ContainsKey(entity.Id))
     {
-      throw new Exception(
+      throw new NoEntityFoundException(
         $"Cannot update {EntityName} with identity '{entity.Id}' as it doesn't exist");
     }
 
@@ -68,5 +81,11 @@ public abstract class InMemoryRepository<T> where T : Entity
     Logger.LogDebug($"New {EntityName} updated [ID:{entity.Id}].");
 
     return Task.FromResult(entity);
+  }
+
+  private void CheckDataInstantiated()
+  {
+    if (Data == null)
+      throw new InvalidRepositoryConfigurationException($"Please ensure that {EntityName} has its Data field instantiated");
   }
 }
