@@ -15,10 +15,10 @@ public class InMemoryCountryRepository : InMemoryRepository<Country>, ICountryRe
     Data = new ConcurrentDictionary<int, Country>();
 
     Data.TryAdd(1, new Country { Id = 1, Name = "Netherlands" });
-    Data.TryAdd(1, new Country { Id = 2, Name = "Poland" });
-    Data.TryAdd(1, new Country { Id = 3, Name = "Ireland" });
-    Data.TryAdd(1, new Country { Id = 4, Name = "South Africa" });
-    Data.TryAdd(1, new Country { Id = 5, Name = "India" });
+    Data.TryAdd(2, new Country { Id = 2, Name = "Poland" });
+    Data.TryAdd(3, new Country { Id = 3, Name = "Ireland" });
+    Data.TryAdd(4, new Country { Id = 4, Name = "South Africa" });
+    Data.TryAdd(5, new Country { Id = 5, Name = "India" });
   }
 
   public Task<IEnumerable<Country>> GetAllAsync()
